Allow favorites on comments

+++ app/controllers/favorites_controller.rb
@@ -3,10 +3,12 @@ class FavoritesController < ApplicationController
     @favorite = Favorite.new(favorite_params)
     @favorite.user_id = current_user.id
 
-    @post = Post.find(favorite_params[:post_id])
+    @model_name = favorite_params[:favoritable_type].constantize
+    return unless defined? @model_name
+    @model = @model_name.find(favorite_params[:favoritable_id])
 
-    if @favorite.save
-      @post.increment!(:hotness)
+    if @favorite.save!
+      @model.increment!(:hotness) if @model_name == "Post"
 
       respond_to do |format|
         format.js
@@ -15,12 +17,14 @@ class FavoritesController < ApplicationController
   end
 
   def destroy
-    @favorite = Favorite.find_by_post_id_and_user_id(favorite_params[:post_id], current_user.id)
+    @favorite = Favorite.find_by_favoritable_type_and_favoritable_id_and_user_id(favorite_params[:favoritable_type], favorite_params[:favoritable_id], current_user.id)
 
-    @post = Post.find(favorite_params[:post_id])
+    @model_name = favorite_params[:favoritable_type].constantize
+    return unless defined? @model_name
+    @model = @model_name.find(favorite_params[:favoritable_id])
 
     if @favorite.destroy
-      @post.increment!(:hotness, -1)
+      @model.increment!(:hotness, -1) if @model_name == "Post"
 
       respond_to do |format|
         format.js
@@ -31,6 +35,6 @@ class FavoritesController < ApplicationController
   private
 
   def favorite_params
-    params.require(:favorite).permit(:post_id)
+    params.require(:favorite).permit(:favoritable_type, :favoritable_id)
   end
 end
